Start of Update 1.4
Added in the world gen for the High Court City, all of its tiles are incomplete as of currently.
Added two new enemies, the High Court Halbidier & Sword Seeker.
Added in some items regarding the structure.
Added a multitude of blocks and tiles.
Added in a new lantern along with a rope for decoration and a lightning piece.