Polish Update Commit One
This updates the Github to 1.5-1.
Completely revamped the Burning Flame Arena in its structure, along with compat with the nether backport that'll allow for different variants to spawn in the nether biomes.

Added a new mob, BareAnt, a small drone like creature found in the Burning Flame Arena.

Revamped the Void Blossom Cave, with a new look and more closer to the original, along with variants of the cave.

Added Particles to the Azalea Vines that add nice ambience.

Added in Boss Enraged System, read up on it in the wiki.

Probably other fixes I forgot about.