1.69 Prep
Finished the entirety of the End Dungeon, including the bosses, mobs, and reward items.

Continued work on the Forgotten Temple. Added the boss room, Aegyptia mob.

Adjusted parts of the mod, particles got a bit of lighting change.

Added a new sound to KoBF's swing to sound more menacing.

Added a new item for Void Blossom to craft.

Added two new blocks that can be found in the Rotten Hold or made via Stonebrick.

Added Damage caps as a new mechanic to all bosses and minibosses

Added more config options.