Mysterious Trader
A new solo villager has been added, the Mysterious Trader. With him you'll be able to exchange all Dungeon Additions mob drops for various other materials or vanilla drops. Ontop of that with coin you'll be to barter practically for new items the Trader offers. With the main selling point of Trinkets.

- Added Mysterious Trader Structure
- Added Mysterious Trader
- Added the trading system
- Added Copper, SIlver, and Gold Coins
- Added 17 new trinkets all with unique effects and conditions for them to be used. (Should spice up building your character a lot)
- Added Novik stuff to config for Gaelon Ingot Repairs.